postProcessing/requirements.txt
nltk==3.7
numpy==1.21.5
pandas==1.3.5
protobuf==4.21.12
scipy==1.7.3
google-cloud-storage==2.7.0
gcsfs==2023.1.0